carInfo 제거
