+++ src/main/java/domain/Car.java
@@ -3,8 +3,9 @@ package domain;
 import model.CarName;
 import model.CarPosition;
 
-public class Car {
+public class Car implements Comparable<Car> {
 
+  private static final String COLUMN_MARKER = " : ";
   private static final int MIN_NUMBER_TO_MOVE = 4;
   private static final int MOVE_DISTANCE = 1;
   private static final int INITIAL_POSITION = 0;
@@ -23,7 +24,20 @@ public class Car {
     }
   }
 
-  public CarInfo getCarInfo() {
-    return new CarInfo(carName, carPosition);
+  public String markPosition() {
+    return carName
+        + COLUMN_MARKER
+        + carPosition;
+  }
+
+
+  @Override
+  public String toString() {
+    return carName.toString();
+  }
+
+  @Override
+  public int compareTo(Car other) {
+    return carPosition.compareTo(other.carPosition);
   }
 }
