+++ src/main/java/model/NumberModel.java
@@ -2,7 +2,7 @@ package model;
 
 public interface NumberModel {
 
-  String toString();
-  int toInt();
+//  String toString();
+//  int toInt();
   void validateValue(int value);
 }
